Edited to completed postorder service with data and transactional

+++ src/main/java/com/example/bookstore/order/OrderRepository.java
@@ -2,5 +2,5 @@ package com.example.bookstore.order;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface OrderRepository extends CrudRepository <UserOrder, Integer> {
+public interface OrderRepository extends CrudRepository <UserOrder, Long> {
 }
